+++ android/app/src/main/assets/translations/ja.json
@@ -59,5 +59,6 @@
   "dtLocal": "各停",
   "rapid": "快速",
   "local": "各駅停車",
-  "ltdExp": "特急"
+  "ltdExp": "特急",
+  "apiErrorText": "大変申し訳ありませんが、リクエストを正常に処理できませんでした。しばらく経ってからもう一度お試しください。"
 }
